+++ backend/db/models/stockusertransaction.js
@@ -16,8 +16,8 @@ module.exports = (sequelize, DataTypes) => {
       userId: { type: DataTypes.INTEGER, allowNull: false },
       stockId: { type: DataTypes.INTEGER, allowNull: false },
       transactionType: { type: DataTypes.STRING, allowNull: false },
-      quantity: { type: DataTypes.INTEGER, allowNull: false },
-      purchasePrice: { type: DataTypes.INTEGER, allowNull: false },
+      quantity: { type: DataTypes.DECIMAL(10,2), allowNull: false },
+      purchasePrice: { type: DataTypes.DECIMAL(10,2), allowNull: false },
       purchaseDate: { type: DataTypes.DATE, allowNull: false },
     },
     {
